Stability Fix (v3.0.1)
When two clickboards used the I2C bus, one of them might lock in a deadlock.
As fix a Mutex for the I2C bus is added. It must be taken before the I2C bus is used in any kind.
MAby the same solution is necessary for other bus systems like SPI or UART, when two clickboards use them simultaneously.

Note: I2C Burst Mode is not supported!