Fix broadcast handling of the bridge
The bridge should forward received broadcast to all interfaces
except the receiving one, which is correct. But the interation
only checked for interfaces of the bridge, and did not forward
these broadcasts to the bridge interface itself and therefore
not to the IP stack, which sits on top of the bridge interface.
This is because the bridge itself is not part of the bridge.

Started the iteration of interfaces with the first network
interface. This works only because the bridge itself is
statically the first interface, so this is more a workaround
than a fix.